site: reallocate requests from users
Users was not the right controller for them, they are generic request
from site.
