+++ site/MVC/Controllers/GenericController.cs
@@ -1,0 +1,23 @@
+﻿using System;
+using System.Text;
+using DFM.Generic;
+using DFM.MVC.Helpers.Controllers;
+using Microsoft.AspNetCore.Mvc;
+
+namespace DFM.MVC.Controllers
+{
+	public class GenericController : Controller
+	{
+		[HttpGetAndHead]
+		public IActionResult Mobile()
+		{
+			return Redirect(Cfg.GooglePlay);
+		}
+
+		public IActionResult Reload()
+		{
+			return Content(" ");
+		}
+
+	}
+}
